requirements.txt
beautifulsoup4==4.14.3
urllib3==1.26.18
python-dotenv==1.2.1
boto3==1.42.59
pandas==2.3.3
pydantic-settings==2.13.1
pymongo==4.16.0
trafilatura==2.0.0
tiktoken==0.12.0
qdrant-client==1.17.0
openai==2.24.0
scikit-learn==1.8.0
FlagEmbedding==1.3.5
deepeval==3.8.9
pyvi==0.0.9.3
flashrank==0.2.10
fastapi==0.135.1
uvicorn==0.42.0
transformers==4.57.6
numpy<2
pwdlib==0.3.0
celery==5.4.0
argon2-cffi==23.1.0
PyJWT==2.10.1
python-multipart==0.0.9
redis==5.0.1
websockets==16.0